+++ apps/ios/Merian/Models/Aliases.swift
@@ -1,4 +1,4 @@
 import SwiftData
 
 // MARK: - Active Schema Aliases
-typealias CurrentSchema = MerianSchemaV46
+typealias CurrentSchema = MerianSchemaV47